refactor: refactor file layout
