+++ src/jsx/dom/intrinsic-element/components.ts
@@ -1,18 +1,18 @@
-import type { Props } from '../base'
-import type { FC, PropsWithChildren, RefObject } from '../types'
-import { newJSXNode } from './utils'
-import { createPortal, getNameSpaceContext } from './render'
-import { useContext } from '../context'
-import { useCallback, useState } from '../hooks'
-import { FormContext } from './hooks'
+import type { Props } from '../../base'
+import type { FC, PropsWithChildren, RefObject } from '../../types'
+import { newJSXNode } from '../utils'
+import { createPortal, getNameSpaceContext } from '../render'
+import { useContext } from '../../context'
+import { useCallback, useState } from '../../hooks'
+import { FormContext } from '../hooks'
 
-const documentMetadataTag = (tag: string, props: Props) => {
+const documentMetadataTag = (tag: string, props: Props, deDupAttrs: string[]) => {
   const jsxNode = newJSXNode({
     tag,
     props,
   })
 
-  if (props?.itemProp || props?.itemprop) {
+  if (props?.itemProp) {
     return jsxNode
   }
 
@@ -44,19 +44,23 @@ export const title: FC<PropsWithChildren> = (props) => {
       props,
     })
   }
-  return documentMetadataTag('title', props)
+  return documentMetadataTag('title', props, [])
 }
 
 export const script: FC<PropsWithChildren> = (props) => {
-  return documentMetadataTag('script', props)
+  return documentMetadataTag('script', props, ['src'])
 }
 
 export const style: FC<PropsWithChildren> = (props) => {
-  return documentMetadataTag('style', props)
+  return documentMetadataTag('style', props, ['href'])
 }
 
 export const link: FC<PropsWithChildren> = (props) => {
-  return documentMetadataTag('link', props)
+  return documentMetadataTag('link', props, ['href'])
+}
+
+export const meta: FC<PropsWithChildren> = (props) => {
+  return documentMetadataTag('meta', props, ['name', 'httpEquiv', 'charset', 'itemProp'])
 }
 
 export const form: FC<
