+++ src/jsx/dom/jsx-dev-runtime.ts
@@ -5,7 +5,7 @@
 
 import type { JSXNode, Props } from '../base'
 import { newJSXNode } from './utils'
-import * as intrinsicElementTags from './intrinsic-element-tags'
+import * as intrinsicElementTags from './intrinsic-element/components'
 
 export const jsxDEV = (tag: string | Function, props: Props, key?: string): JSXNode => {
   return newJSXNode({
