+++ src/jsx/intrinsic-element/components.ts
@@ -1,10 +1,10 @@
-import type { HtmlEscapedCallback, HtmlEscapedString } from '../utils/html'
-import { JSXNode } from './base'
-import type { Child, Props } from './base'
-import type { FC, PropsWithChildren } from './types'
-import { raw } from '../helper/html'
+import type { HtmlEscapedCallback, HtmlEscapedString } from '../../utils/html'
+import { JSXNode } from '../base'
+import type { Child, Props } from '../base'
+import type { FC, PropsWithChildren } from '../types'
+import { raw } from '../../helper/html'
 
-const metaTagMap: WeakMap<Object, Record<string, [string, string | undefined][]>> = new WeakMap()
+const metaTagMap: WeakMap<object, Record<string, [string, string | undefined][]>> = new WeakMap()
 const insertIntoHead: (
   tagName: string,
   tag: string,
@@ -53,7 +53,7 @@ const documentMetadataTag = (tag: string, children: Child, props: Props, sort: b
 
   const string = new JSXNode(tag, props, children as Child[]).toString()
 
-  if (props?.itemProp || props?.itemprop) {
+  if (props?.itemProp) {
     return raw(string)
   }
 
@@ -80,6 +80,9 @@ export const style: FC<PropsWithChildren> = ({ children, ...props }) => {
 }
 export const link: FC<PropsWithChildren> = ({ children, ...props }) => {
   return documentMetadataTag('link', children, props, true)
+}
+export const meta: FC<PropsWithChildren> = ({ children, ...props }) => {
+  return documentMetadataTag('meta', children, props, true)
 }
 export const form: FC<
   PropsWithChildren<{
